Module google_servicecontrol1::api

Structs

AllocateInfo
There is no detailed description.

AllocateQuotaRequest
Request message for the AllocateQuota method.

AllocateQuotaResponse
Response message for the AllocateQuota method.

AttributeValue
The allowed types for [VALUE] in a [KEY]:[VALUE] attribute.

Attributes
A set of attributes, each in the format [KEY]:[VALUE].

CheckError
Defines the errors to be returned in google.api.servicecontrol.v1.CheckResponse.check_errors.

CheckInfo
Contains additional information about the check operation.

CheckRequest
Request message for the Check method.

CheckResponse
Response message for the Check method.

ConsumerInfo
ConsumerInfo provides information about the consumer.

Distribution
Distribution represents a frequency distribution of double-valued sample points. It contains the size of the population of sample points plus additional optional information: - the arithmetic mean of the samples - the minimum and maximum of the samples - the sum-squared-deviation of the samples, used to compute variance - a histogram of the values of the sample points

Exemplar
Exemplars are example points that may be used to annotate aggregated distribution values. They are metadata that gives information about a particular value added to a Distribution bucket, such as a trace ID that was active when a value was added. They may contain further information, such as a example values and timestamps, origin, etc.

ExplicitBuckets
Describing buckets with arbitrary user-provided width.

ExponentialBuckets
Describing buckets with exponentially growing width.

HttpRequest
A common proto for logging HTTP requests. Only contains semantics defined by the HTTP specification. Product-specific logging information MUST be defined in a separate message.

LinearBuckets
Describing buckets with constant width.

LogEntry
An individual log entry.

LogEntryOperation
Additional information about a potentially long-running operation with which a log entry is associated.

LogEntrySourceLocation
Additional information about the source code location that produced the log entry.

MetricValue
Represents a single metric value.

MetricValueSet
Represents a set of metric values in the same metric. Each metric value in the set should have a unique combination of start time, end time, and label values.

Money
Represents an amount of money with its currency type.

Operation
Represents information regarding an operation.

QuotaError
Represents error information for QuotaOperation.

QuotaInfo
Contains the quota information for a quota check response.

QuotaOperation
Represents information regarding a quota operation.

QuotaProperties
Represents the properties needed for quota operations.

ReportError
Represents the processing error of one Operation in the request.

ReportRequest
Request message for the Report method.

ReportResponse
Response message for the Report method.

ResourceInfo
Describes a resource associated with this operation.

ServiceAllocateQuotaCall
Attempts to allocate quota for the specified consumer. It should be called before the operation is executed. This method requires the servicemanagement.services.quota permission on the specified service. For more information, see Cloud IAM. NOTE: The client must fail-open on server errors INTERNAL, UNKNOWN, DEADLINE_EXCEEDED, and UNAVAILABLE. To ensure system reliability, the server may inject these errors to prohibit any hard dependency on the quota functionality.

ServiceCheckCall
Checks whether an operation on a service should be allowed to proceed based on the configuration of the service and related policies. It must be called before the operation is executed. If feasible, the client should cache the check results and reuse them for 60 seconds. In case of any server errors, the client should rely on the cached results for much longer time to avoid outage. WARNING: There is general 60s delay for the configuration and policy propagation, therefore callers MUST NOT depend on the Check method having the latest policy information. NOTE: the CheckRequest has the size limit of 64KB. This method requires the servicemanagement.services.check permission on the specified service. For more information, see Cloud IAM.

ServiceControl
Central instance to access all ServiceControl related resource activities

ServiceMethods
A builder providing access to all methods supported on service resources. It is not used directly, but through the ServiceControl hub.

ServiceReportCall
Reports operation results to Google Service Control, such as logs and metrics. It should be called after an operation is completed. If feasible, the client should aggregate reporting data for up to 5 seconds to reduce API traffic. Limiting aggregation to 5 seconds is to reduce data loss during client crashes. Clients should carefully choose the aggregation time window to avoid data loss risk more than 0.01% for business and compliance reasons. NOTE: the ReportRequest has the size limit (wire-format byte size) of 1MB. This method requires the servicemanagement.services.report permission on the specified service. For more information, see Google Cloud IAM.

Status
The Status type defines a logical error model that is suitable for different programming environments, including REST APIs and RPC APIs. It is used by gRPC. Each Status message contains three pieces of data: error code, error message, and error details. You can find out more about this error model and how to work with it in the API Design Guide.

TraceSpan
A span represents a single operation within a trace. Spans can be nested to form a trace tree. Often, a trace contains a root span that describes the end-to-end latency, and one or more subspans for its sub-operations. A trace can also contain multiple root spans, or none at all. Spans do not need to be contiguous—there may be gaps or overlaps between spans in a trace.

TruncatableString
Represents a string that might be shortened to a specified length.

Enums

Scope
Identifies the an OAuth2 authorization scope. A scope is needed when requesting an authorization token.